Move intervetionProtocol to intervention module
